+++ src/components/ExpertsList.vue
@@ -27,7 +27,7 @@
       </tbody>
     </table>
   </div>
-  <ContextMenu v-if="showMenu" :on-close="closeContextMenu" :actions="contextMenuActions" @action-clicked="handleActionClick" :x="menuX" :y="menuY" position="right" :teleport="false" />
+  <ContextMenu v-if="showMenu" @close="closeContextMenu" :actions="contextMenuActions" @action-clicked="handleActionClick" :x="menuX" :y="menuY" position="right" :teleport="false" />
 </template>
 
 <script setup lang="ts">