+++ src/main/window.ts
@@ -4,6 +4,7 @@ export * from './windows/index';
 export * from './windows/main';
 export * from './windows/anywhere';
 export * from './windows/commands';
+export * from './windows/forge';
 export * from './windows/settings';
 export * from './windows/readaloud';
 export * from './windows/realtime';