+++ src/models/agent.ts
@@ -9,6 +9,7 @@ export default class Agent implements AgentBase {
   updatedAt: number
   name: string
   description: string
+  primary: boolean
   engine: string
   model: string
   locale: string
@@ -28,6 +29,7 @@ export default class Agent implements AgentBase {
     this.updatedAt = Date.now()
     this.name = ''
     this.description = ''
+    this.primary = true
     this.engine = ''
     this.model = ''
     this.modelOpts = {}
@@ -46,7 +48,8 @@ export default class Agent implements AgentBase {
     obj: any,
     preparationDescription?: () => string,
     runningDescription?: (args: any) => string,
-    completedDescription?: (args: any, results: any) => string
+    completedDescription?: (args: any, results: any) => string,
+    errorDescription?: (args: any, results: any) => string
   ): Agent {
     const agent = new Agent()
     agent.id = obj.id || crypto.randomUUID()
@@ -54,6 +57,7 @@ export default class Agent implements AgentBase {
     agent.updatedAt = obj.updatedAt ?? Date.now()
     agent.name = obj.name
     agent.description = obj.description
+    agent.primary = obj.primary ?? true
     agent.engine = obj.engine ?? ''
     agent.model = obj.model ?? ''
     agent.modelOpts = obj.modelOpts ?? null
@@ -69,6 +73,7 @@ export default class Agent implements AgentBase {
     agent.getPreparationDescription = preparationDescription
     agent.getRunningDescription = runningDescription
     agent.getCompletedDescription = completedDescription
+    agent.getErrorDescription = errorDescription
     return agent
   }
 
@@ -76,7 +81,10 @@ export default class Agent implements AgentBase {
     if (!this.prompt) return null
     let prompt = this.prompt
     for (const param of Object.keys(parameters)) {
-      const value = parameters[param]
+      let value = parameters[param]
+      if (Array.isArray(value)) {
+        value = value.join(', ')
+      }
       prompt = prompt.replace(new RegExp(`{{${param}}}`, 'g'), value)
     }
     return prompt
@@ -85,5 +93,6 @@ export default class Agent implements AgentBase {
   getPreparationDescription?: () => string
   getRunningDescription?: (args: any) => string
   getCompletedDescription?: (args: any, results: any) => string
+  getErrorDescription?: (args: any, results: any) => string
 
 }