+++ CHANGELOG.md
@@ -2,16 +2,39 @@
 
 All notable changes to this project will be documented in this file.
 
-## [2.10.0] - YYYY-MM-DD
+## [2.11.0] - 2025-07-07
+
+### Added
+- Custom HTTP Headers for MCP Streamable
+- File upload for transcriptions (with dropzone)
+- Summarize/Translate/Run AI command for transcription
+- Drag and drop to attach files
+
+### Changed
+- N/A 
+
+### Fixed
+- N/A 
+
+### Removed
+- N/A
+
+
+
+## [2.10.0] - 2025-07-03
 
 ### Added
 - DeepResearch
+- Fileystem plugin to read/write local files
 
 ### Changed
-- N/A
+- Text headings font size and spacing 
 
 ### Fixed
 - PDF export when tools displayed
+- Fullscreen exit requiring multiple clicks
+- YouTube transcript download
+- Duplicate MCP servers sent to model ([#302](https://github.com/nbonamy/witsy/issues/302))
 
 ### Removed
 - N/A